requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-text-splitters==0.2.2
sentence-transformers==3.0.1
pypdf==4.3.1
python-dotenv==1.0.1
streamlit==1.37.0
huggingface_hub==0.24.5
#faiss-cpu==1.8.0.post1

